Group by takes into account filters, now keeping headers names in meta data and h2 in sync
git-svn-id: svn://107.170.10.188:8080/Semoss/dev@13683 ef391c1c-ee15-45f7-b1d6-dde3d27a2169

Former-commit-id: 3fc8c51bc26434cb2b321aa5136c0e69252a9ccc

+++ src/prerna/ds/H2/TinkerH2Frame.java
@@ -344,7 +344,22 @@ public class TinkerH2Frame extends TinkerFrame {
 	
 	@Override
 	public void removeColumn(String columnHeader) {
+		if(!ArrayUtilityMethods.arrayContainsValue(this.headerNames, columnHeader)) {
+			return;
+		}
+		
 		builder.dropColumn(columnHeader);
+		
+		String[] newHeaders = new String[this.headerNames.length-1];
+		int newHeaderIdx = 0;
+		for(int i = 0; i < this.headerNames.length; i++){
+			String name = this.headerNames[i];
+			if(!name.equals(columnHeader)){
+				newHeaders[newHeaderIdx] = name;
+				newHeaderIdx ++;
+			}
+		}
+		this.headerNames = newHeaders;
 	}
 	
 	@Override
